requirements.txt
Flask
Flask-SQLAlchemy
flask-bcrypt 
Flask-JWT-Extended
werkzeug
flask-login
python-dotenv
